Refactor EditCommand to EditCustomerCommand
