+++ src/test/java/seedu/address/logic/commands/EditCommandStub.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands;
 
-public class EditCommandStub extends EditCommand {
-    public EditCommandStub(EditCommand editCommand, EditCustomerDescriptor editCustomerDescriptor) {
+public class EditCommandStub extends EditCustomerCommand {
+    public EditCommandStub(EditCustomerCommand editCommand, EditCustomerDescriptor editCustomerDescriptor) {
         super(editCommand.index, editCustomerDescriptor);
     }
 }
