+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -25,8 +25,8 @@ import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.DeleteCheeseCommand;
 import seedu.address.logic.commands.DeleteCustomerCommand;
 import seedu.address.logic.commands.DeleteOrderCommand;
-import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.EditCommandStub;
+import seedu.address.logic.commands.EditCustomerCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCustomerCommand;
 import seedu.address.logic.commands.HelpCommand;
@@ -108,11 +108,11 @@ public class AddressBookParserTest {
     @Test
     public void parseCommand_edit() throws Exception {
         Customer customer = new CustomerBuilder().withId(CustomerIdStub.getNextId()).build();
-        EditCommand.EditCustomerDescriptor descriptor = new EditCustomerDescriptorBuilder(customer).build();
-        EditCommand command = (EditCommand) parser.parseCommand(EditCommand.COMMAND_WORD + " "
+        EditCustomerCommand.EditCustomerDescriptor descriptor = new EditCustomerDescriptorBuilder(customer).build();
+        EditCustomerCommand command = (EditCustomerCommand) parser.parseCommand(EditCustomerCommand.COMMAND_WORD + " "
                 + INDEX_FIRST_CUSTOMER.getOneBased() + " " + CustomerUtil.getEditCustomerDescriptorDetails(descriptor));
         command = new EditCommandStub(command, descriptor);
-        assertEquals(new EditCommand(INDEX_FIRST_CUSTOMER, descriptor), command);
+        assertEquals(new EditCustomerCommand(INDEX_FIRST_CUSTOMER, descriptor), command);
     }
 
     @Test
