+++ src/test/java/seedu/address/testutil/CustomerUtil.java
@@ -9,7 +9,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 import java.util.Set;
 
 import seedu.address.logic.commands.AddCustomerCommand;
-import seedu.address.logic.commands.EditCommand;
+import seedu.address.logic.commands.EditCustomerCommand;
 import seedu.address.model.customer.Customer;
 import seedu.address.model.tag.Tag;
 
@@ -43,7 +43,7 @@ public class CustomerUtil {
     /**
      * Returns the part of command string for the given {@code EditCustomerDescriptor}'s details.
      */
-    public static String getEditCustomerDescriptorDetails(EditCommand.EditCustomerDescriptor descriptor) {
+    public static String getEditCustomerDescriptorDetails(EditCustomerCommand.EditCustomerDescriptor descriptor) {
         StringBuilder sb = new StringBuilder();
         descriptor.getName().ifPresent(name -> sb.append(PREFIX_NAME).append(name.fullName).append(" "));
         descriptor.getPhone().ifPresent(phone -> sb.append(PREFIX_PHONE).append(phone.value).append(" "));
